improved code transparency
