+++ budget_manager/views.py
@@ -1,5 +1,5 @@
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.db.models import Sum, Subquery
+from django.db.models import Sum
 from django.db.models.functions import ExtractMonth
 from django.views.generic.base import TemplateView
 from django.views.generic.edit import CreateView, UpdateView, DeleteView
@@ -7,69 +7,32 @@ from django.views.generic.list import ListView
 from django_filters.views import FilterView
 from django.urls import reverse_lazy
 from django.contrib.auth.models import User
-from budget_manager.forms import CurrencyFilter,UpdateUserForm
 
+from .utils import create_plt
+from budget_manager.forms import CurrencyFilter, UpdateUserForm
 from .filtersets import ExpenseFilter, IncomeFilter
-import matplotlib.pyplot as plt
 from budget_manager.models import Expense, Source, Category, Income, Currency
-
-
-def create_plt(*args):
-    try:
-        fig = plt.figure(figsize=(8, 5))
-        plt.bar(['Wydatki', 'Przychody'], args, color=['red', 'green'], width=0.4)
-        plt.xlabel('Rodzaj')
-        plt.ylabel('Wartość')
-        plt.title('Bilans finansów')
-        plt.savefig('budget_manager/static/budget_manager/expense.jpg')
-    except TypeError:
-        print('Brak danych')
 
 
 class BalanceView(LoginRequiredMixin, TemplateView):
     template_name = 'home.html'
 
-    # def get(self, request, *args, **kwargs):
-    #     period = self.request.GET.get('period', '')
-    def get(self, request, *args, **kwargs):
-        context = self.get_context_data(**kwargs)
-        if context["currency_filter_form"].is_valid():
-            print("Valid get form!!")
-        return self.render_to_response(context)
-
     def get_context_data(self, **kwargs):
         currency_filter_form = CurrencyFilter(self.request.GET)
-        if currency_filter_form.data.__len__() is 0:
-            pass
+        if currency_filter_form.data.__len__() == 0:
+            filtered_currency_pk = 1
         else:
-            filtred_currency_pk = currency_filter_form.data["currency_filter"]
-            print(filtred_currency_pk)
+            filtered_currency_pk = currency_filter_form.data["currency_filter"]
 
-        # test = if currency_filter_form.data else None
-        # print()
-        # print(filtred_currency_pk)
-        # print(currency_filter_form)
         total_expense = round(Expense.objects.filter(user=self.request.user).filter(
-            currency__id=1).aggregate(Sum("cost", default=0))["cost__sum"], ndigits=2) \
-            if  currency_filter_form.data.__len__() is 0 \
-            else round(Expense.objects.filter(user=self.request.user).filter(
-            currency__id=filtred_currency_pk).aggregate(Sum("cost", default=0))["cost__sum"], ndigits=2)
-
+            currency__id=filtered_currency_pk).aggregate(Sum("cost", default=0))["cost__sum"], ndigits=2)
         total_income = round(Income.objects.filter(user=self.request.user).filter(
-            currency__id=1).aggregate(Sum("amount", default=0))["amount__sum"], ndigits=2) \
-            if currency_filter_form.data.__len__() is 0 \
-            else round(Income.objects.filter(user=self.request.user).filter(
-            currency__id=filtred_currency_pk).aggregate(Sum("amount", default=0))["amount__sum"], ndigits=2)
-
+            currency__id=filtered_currency_pk).aggregate(Sum("amount", default=0))["amount__sum"], ndigits=2)
         total_balance = round(total_income - total_expense, ndigits=2)
-
         monthly_income = Income.objects.filter(user=self.request.user).annotate(
             month=ExtractMonth('income_date')).values('month').annotate(total_amount=Sum('amount'))
-        print(monthly_income)
         monthly_expense = Expense.objects.filter(user=self.request.user).annotate(
             month=ExtractMonth('expense_date')).values('month').annotate(total_amount=Sum('cost'))
-
-
 
         context = {
             'total_expense': total_expense,
@@ -78,12 +41,11 @@ class BalanceView(LoginRequiredMixin, TemplateView):
             'monthly_income': monthly_income,
             'monthly_expense': monthly_expense,
             "currency_filter_form": currency_filter_form,
-            "displayed_currency": Currency.objects.get(pk=1 if currency_filter_form.data.__len__() is 0 \
-                else filtred_currency_pk)
+            "displayed_currency": Currency.objects.get(pk=filtered_currency_pk)
         }
+
         create_plt(total_expense, total_income)
-        print(dir(monthly_income))
-        print(monthly_income.values)
+
         return context
 
 
@@ -102,8 +64,8 @@ class ExpenseCreateView(LoginRequiredMixin, CreateView):
     def get_form(self, *args, **kwargs):
         """Displays only categories made by currently logged user"""
         form = super(ExpenseCreateView, self).get_form(*args, **kwargs)
-        form.fields['category'].queryset = Category.objects.filter(user=self.request.user) | \
-                                           Category.objects.filter(user=None)
+        form.fields['category'].queryset = Category.objects.filter(
+            user=self.request.user) | Category.objects.filter(user=None)
         field_labels = {'name': 'Nazwa',
                         'cost': 'Koszt',
                         'expense_date': 'Data wydatku',
@@ -181,8 +143,8 @@ class IncomeCreateView(LoginRequiredMixin, CreateView):
     def get_form(self, *args, **kwargs):
         """Displays only categories made by currently logged user"""
         form = super(IncomeCreateView, self).get_form(*args, **kwargs)
-        form.fields['source'].queryset = Source.objects.filter(user=self.request.user) | \
-                                         Source.objects.filter(user=None)
+        form.fields['source'].queryset = Source.objects.filter(
+            user=self.request.user) | Source.objects.filter(user=None)
 
         field_labels = {'amount': 'Wartość',
                         'income_date': 'Data przychodu',
